fix: 게시글 정보 불러오기 무한 스크롤 => 타입 에러 수정
