+++ src/queries/post/getPosts.ts
@@ -1,28 +1,7 @@
 'use server';
 
+import PostListType from '@ts/postType';
 import { createClient } from '@utils/supabase/server';
-
-import { Database } from '@ts/supabase';
-
-export type PostType = Pick<Database['public']['Tables']['posts']['Row'], 'id' | 'content'>;
-export type GroupType = Pick<Database['public']['Tables']['groups']['Row'], 'name' | 'description'>;
-export type UserType = Pick<Database['public']['Tables']['users']['Row'], 'id' | 'nickname' | 'profile_image'>;
-export type ScheduleType = Pick<
-  Database['public']['Tables']['schedules']['Row'],
-  'name' | 'memo' | 'start_date' | 'end_date' | 'start_time'
->;
-export type CommentCountType = { count: number };
-export type PostImageType = { image_url: string };
-
-export type PostListType = {
-  id: PostType['id'];
-  content: PostType['content'];
-  groups: GroupType;
-  users: UserType;
-  schedules: ScheduleType;
-  comments: CommentCountType;
-  post_images: PostImageType[];
-};
 
 // 게시글 리스트 불러오기
 export const getPosts = async (groupId: string, offset: number, limit: number): Promise<PostListType[]> => {
